+++ crates/rvr/rvr-openvm/c/tracer/openvm_tracer_pure.h
@@ -68,12 +68,12 @@ static __attribute__((always_inline)) inline void trace_mem_access_u64_range(
     uint32_t addr_space) {}
 
 static __attribute__((always_inline)) inline void trace_pc(
-    RvState* restrict state, uint32_t pc) {}
+    RvState* restrict state, uint64_t pc) {}
 
 static __attribute__((always_inline)) inline void trace_chip(
     RvState* restrict state, uint32_t chip_idx, uint32_t count) {}
 
 static __attribute__((always_inline)) inline void trace_block(
-    RvState* restrict state, uint32_t pc, uint32_t block_insn_count) {}
+    RvState* restrict state, uint64_t pc, uint32_t block_insn_count) {}
 
 #endif /* OPENVM_TRACER_PURE_H */
